==> 2_icd_processing.ipynb <==
import pandas as pd
from google.colab import drive
drive.mount('/content/drive')

## Process ICD-10 files into one

# load in icd files
## files were obtained from: https://www.cdc.gov/nchs/icd/icd-10-cm/files.html
icd10_2015 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2015.csv')
icd10_2016 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2016.csv')
icd10_2017 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2017.csv')
icd10_2018 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2018.csv')
icd10_2019 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2019.csv')
icd10_2020 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2020.csv')
icd10_2021 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2021.csv')
icd10_2022 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2022.csv')
icd10_2023 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2023.csv')
icd10_2024 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2024.csv')
icd10_2025 = pd.read_csv('/content/drive/MyDrive/Data/icd10/icd10-2025.csv')

icd10_joined = pd.concat([icd10_2025, icd10_2024, icd10_2023, icd10_2022, icd10_2021, icd10_2020,
                          icd10_2019, icd10_2018, icd10_2017, icd10_2016, icd10_2015])

# prompt: drop all duplicates from icd10_joined according to the variable "Code"

icd10_joined = icd10_joined.drop_duplicates(subset=['Code'])

icd10_joined.to_csv('/content/drive/MyDrive/Data/icd10/ICD10Diagnosis.csv', index=False)

## Process CCSR look up file

# CCSR lookup file for ICD10 regrouping
## look up was obtained from HCUP: https://hcup-us.ahrq.gov/toolssoftware/ccsr/dxccsr.jsp#download
LU_CCSR = pd.read_csv('/content/drive/MyDrive/Data/Input/DXCCSR_v2025-1.csv')
LU_CCSR.head()

# Remove '' from cells and column names in LU_CCSR
LU_CCSR.columns = LU_CCSR.columns.str.strip("'")
for column in LU_CCSR.columns:
  if LU_CCSR[column].dtype == object:
    LU_CCSR[column] = LU_CCSR[column].str.strip("'")

# select needed columns, then rename
LU_CCSR = LU_CCSR[['ICD-10-CM CODE', 'CCSR CATEGORY 1', 'CCSR CATEGORY 1 DESCRIPTION']]
LU_CCSR.rename(columns={'ICD-10-CM CODE':'ICD_CD', 'CCSR CATEGORY 1':'CCSR_CATEGORY_CD','CCSR CATEGORY 1 DESCRIPTION': 'CCSR_CATEGORY_DESCRIPTION'}, inplace=True)
LU_CCSR

LU_CCSR.to_csv('/content/drive/MyDrive/Data/Output/DXCCSR_v2025-1_processed.csv', index=False)
