==> 5b_diabetes_progression_transformation.ipynb <==
import pandas as pd
import numpy as np
from google.colab import drive
drive.mount('/content/drive')

### Compare structure of raw data and transformed data

# read untransformed data
diabetes_cohort = pd.read_csv('/content/drive/MyDrive/Data/Output/diabetes_cohort.csv')
diabetes_cohort.head(7)

# read transformed data
diabetes_cohort_transformed = pd.read_csv('/content/drive/MyDrive/Data/Output/diab_cohort.csv')
diabetes_cohort_transformed
