==> 6_modeling.Rmd <==
---
title: 'Modeling'
author: ""
date: ""
output:
  html_document:
    df_print: paged
    toc: TRUE
    toc_float: TRUE
    number_sections: TRUE
classoption: landscape
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
library(here)
library(dplyr)
library(tidyr)
library(tibble)
library(survival)
library(survminer)
library(cowplot)
library(gt)
library(msm)
library(ggplot2)
```



```{r}
data <- readRDS(file=here("diab_cohort.rds"))
data
```



## Kaplan Meier 

* Here, we attempt to estimate survival probabilities; the probability that beneficiary has not yet made transition at any given time.

* For example, for a beneficiary with prediabetes (`state=1`), we want to estimate the probability that at time, say, 6 months, they still had prediabetes and hadn't transitioned to any higher state.

```{r}
# time of first appearance in state 1 and time of next transition to state 2/3
survival_1 <- data %>%
  group_by(BENE_ID) %>%
  arrange(TIME_SINCE_PREDIAB) %>%
  filter(STATE %in% c(1, 2, 3)) %>%
  summarize(
    # (1) time at state=1 (value is 0 for all beneficiaries)
    # (2) find first/earliest time beneficiary transitions to state 2 or 3
    ##... ie., when did the beneficiary progress to either diabetes without/with complications?
    # (3) which state did beneficiary progress to ?
    start_time = TIME_SINCE_PREDIAB[which(STATE == 1)[1]],
    next_time = TIME_SINCE_PREDIAB[which(STATE %in% c(2, 3))[1]],
    next_state = STATE[which(STATE %in% c(2, 3))[which(STATE %in% c(2, 3)) > which(STATE == 1)[1]][1]]) %>%
  mutate(time_to_event = next_time - start_time,
         event = ifelse(!is.na(next_state), 1, 0)) %>%
  mutate(time_to_event = ifelse(is.na(time_to_event), max(time_to_event, na.rm=TRUE), time_to_event)) %>%
  arrange(desc(BENE_ID))

survival_1
```


```{r}
# time of first appearance in state 2 and time of transition to state 3
survival_2 <- data %>%
  group_by(BENE_ID) %>%
  arrange(TIME_SINCE_PREDIAB) %>%
  filter(STATE %in% c(2, 3)) %>%
  summarize(
    # (1) time at state=2 
    # (2) find first/earliest time beneficiary transitions to state 3
    ##... ie., when did the beneficiary progress to diabetes with complications?
    start_time = TIME_SINCE_PREDIAB[which(STATE == 2)[1]],
    next_time = TIME_SINCE_PREDIAB[which(STATE == 3)[1]],
    next_state = STATE[which(STATE == 3)[which(STATE == 3) > which(STATE == 2)[1]][1]]) %>%
  mutate(time_to_event = next_time - start_time,
         event = ifelse(!is.na(next_state), 1, 0)) %>%
  filter(!is.na(start_time)) %>% #filter out patients who never entered state 2
  mutate(time_to_event = ifelse(is.na(time_to_event), max(time_to_event, na.rm=TRUE), time_to_event)) %>%
  arrange(desc(BENE_ID))

survival_2
```


```{r}
# K-M curves
km_1 <- survfit(Surv(time_to_event, event) ~ 1, data = survival_1)
km_2 <- survfit(Surv(time_to_event, event) ~ 1, data = survival_2)

kmplot_1 <- ggsurvplot(
  km_1, conf.int = TRUE, surv.median.line = "hv", xlab = "Months since Prediabetes")$plot + 
  theme(legend.position = "none") +
  scale_y_continuous(limits = c(0.55, 1))

kmplot_2 <- ggsurvplot(
  km_2, conf.int = TRUE, surv.median.line = "hv", xlab = "Months since Diabetes wo-comp.")$plot + 
  theme(legend.position = "none") +
  scale_y_continuous(limits = c(0.55, 1))

combined_plot <- plot_grid(kmplot_1, kmplot_2, labels = c("A", "B"), ncol = 2)
combined_plot
```


## Multi-state Modeling

### Transition Probabilities

* Here, we estimate transition probabilities; the probability of being in a future state (or remain in same state) at given times, given that the patient was in that starting state.

* For example, we want to estimate the probability that a patient is with diabetes with complications (`state=3`) at time, say, 6 months, given that they had prediabetes at baseline.

$$P_{ij}(t) = Pr(\text{patient in state}\,j\,\text{at time}\,t\, |\,\text{patient started in state}\,i)$$
* We'll compute these probabilities while adjusting for demographic factors (sex, race, age at baseline) and comorbidities (Obesity, Socioeconomic or lifestyle factors, Hypertension, Heart disease or failure, Sleep wake disorders, Alcohol-related disorders, Depressive or anxiety disorders).

* Refer to `R` package `msm` for fitting continuous-time **multi-state models**.


```{r}
Q <- rbind(c(0, 0.1, 0.1), # 1 to 2/3
           c(0, 0,   0.1), # 2 to 3
           c(0, 0,   0)) # 3 (absorbing state)

model <- msm(
  STATE ~ TIME_SINCE_PREDIAB, subject = BENE_ID,
  data = data, qmatrix = Q,
  covariates = ~ Obesity + `Socioeconomic or lifestyle factors` + Hypertension +
    `Heart disease or failure` +  `Sleep wake disorders` + `Alcohol-related disorders` +
    `Depressive or anxiety disorders` + SEX_IDENT_CD + BENE_RACE_CD + AGE_CALC,
  method = "BFGS")

pmatrix.msm(model, t=5)
```



* Obtain transition probabilities for first 12 months and other later months
* These probabilities are constrained on all other covariates being unobserved, eg., `Obesity=0`, `Socioeconomic or lifestyle factors=0`,etc.

```{r}
times <- c(0:12, seq(15, max(data$TIME_SINCE_PREDIAB)+3, 3))

# transition probabilities
transition_state1 <- transition_state2 <- c()

for (tm in times) {
  transition_state1 <- rbind(transition_state1, as.vector(pmatrix.msm(model, t = tm)[1,]))
  transition_state2 <- rbind(transition_state2, as.vector(pmatrix.msm(model, t = tm)[2,2:3]))
}

colnames(transition_state1) <- c("State 1","State 2","State 3")
colnames(transition_state2) <- c("State 2","State 3")
rownames(transition_state1) <- rownames(transition_state2) <- times
#transition_state1
#transition_state2

tbl1 <- round(transition_state1, 3) %>%
  as.data.frame() %>%
  rownames_to_column(var = "Time (months)") %>%
  gt() %>%
  tab_source_note(md("Abbreviations: State 1 = Prediabetes, 
                     State 2 = Diabetes mellitus without complication,
                     State 3 = Diabetes mellitus with complication")) %>%
  tab_header(md("**Transition Probabilities from Prediabetes**"))

tbl1

tbl2 <- round(transition_state2, 3) %>% 
  as.data.frame(transition_state2) %>%
  rownames_to_column(var = "Time (months)") %>%
  gt()%>%
  tab_source_note(md("Abbreviations: State 1 = Prediabetes, 
                     State 2 = Diabetes mellitus without complication,
                     State 3 = Diabetes mellitus with complication")) %>%
  tab_header(md("**Transition Probabilities from Diabetes without Complications**"))

tbl2
```



```{r}
# transition probability curves
transition_state1_curve <- as.data.frame(transition_state1) %>%
  rownames_to_column(var = "time") %>%
  mutate(trns_from = "State 1")

transition_state2_curve <- as.data.frame(transition_state2) %>%
  rownames_to_column(var = "time") %>%
  mutate(`State 1` = NA, trns_from = "State 2") %>%
  select(time, `State 1`, `State 2`, `State 3`, trns_from) 

transition_curve <- rbind(transition_state1_curve, transition_state2_curve) %>%
  mutate(time = as.numeric(time)) %>%
  pivot_longer(cols = starts_with("State"), names_to = "State", values_to = "prob") %>%
  mutate(State = stringr::str_sub(State, 7)) %>%
  ggplot(., aes(x = time, y = prob, color = State)) +
  geom_line() +
  facet_wrap(trns_from ~., nrow = 1) +
  labs(x="Time (months)", y="Probability", title="Transition Probabilities Over Time") +
  theme_pubr()

transition_curve
```



### Covariate Effects across Diabetes States

```{r}
# get HRs and CIs
hrs <- hazard.msm(model)
hrs$Obesity

## just a little function to summarize HR output into format: HR (L, U)
# hr_format <- function(covariate) {
#   apply(covariate, 1, function(x) {
#     x = round(x, 3); return(paste0(x[1], " (", x[2], ", ", x[3], ")"))})}

hr_format <- function(covariate) {
  apply(covariate, 1, function(x, M=1e+6) {
    x <- round(x, 3)
    if (x[3] < M) {return(paste0(x[1], " (", x[2], ", ", x[3], ")"))}
    else {return(paste0(x[1], " (", x[2], ", ", Inf, ")"))}
    })}

hrs <- lapply(hrs, hr_format)
hr_summary <- do.call(rbind, hrs)
colnames(hr_summary) <- c("1-2","1-3","2-3")
rownames(hr_summary) <- c(names(data)[4:10], "Sex", "Race", "Age (years)")

hr_summary <- as.data.frame(hr_summary) %>%
  rownames_to_column(var = "Covariate") %>%
  gt()%>%
  tab_source_note(md("1 = Prediabetes, 
                     2 = Diabetes mellitus without complication,
                     3 = Diabetes mellitus with complication")) %>%
  tab_header(md("**Estimated Effect per Transition**"))

hr_summary
gtsave(data = hr_summary, filename =  "hr_summary.html", path = here::here())
```
