==> 7_subgroup_analysis.Rmd <==
---
title: 'Subgroup Analysis'
author: ""
date: ""
output:
  html_document:
    df_print: paged
    toc: TRUE
    toc_float: TRUE
    number_sections: TRUE
classoption: landscape
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE)
library(here)
library(dplyr)
library(tidyr)
library(tibble)
library(survival)
library(survminer)
library(cowplot)
library(gt)
library(msm)
library(ggplot2)

data <- readRDS(file=here("diab_cohort.rds"))

## just a little function to summarize HR output into format: HR (L, U)
hr_format <- function(covariate) {
  apply(covariate, 1, function(x, M=1e+6) {
    x <- round(x, 3)
    if (x[3] < M) {return(paste0(x[1], " (", x[2], ", ", x[3], ")"))}
    else {return(paste0(x[1], " (", x[2], ", ", Inf, ")"))}
    })}
```



## Statify by Sex
* Let's compare sex-adjusted transition probabilities between the two sexes.
* To explore this transition dynamics, we refit separate MSM for each sex group

```{r}
table(data$SEX_IDENT_CD)
```


```{r}
Q <- rbind(c(0, 0.1, 0.1), # 1 to 2/3
           c(0, 0,   0.1), # 2 to 3
           c(0, 0,   0)) # 3 (absorbing state)

# separate msm for males vs females
model_male <- msm(
  STATE ~ TIME_SINCE_PREDIAB, subject = BENE_ID,
  data = subset(data, SEX_IDENT_CD==1), qmatrix = Q,
  covariates = ~ Obesity + `Socioeconomic or lifestyle factors` + Hypertension +
    `Heart disease or failure` +  `Sleep wake disorders` + `Alcohol-related disorders` +
    `Depressive or anxiety disorders` + BENE_RACE_CD + AGE_CALC,
  method = "BFGS")

model_female <- msm(
  STATE ~ TIME_SINCE_PREDIAB, subject = BENE_ID,
  data = subset(data, SEX_IDENT_CD==2), qmatrix = Q,
  covariates = ~ Obesity + `Socioeconomic or lifestyle factors` + Hypertension +
    `Heart disease or failure` +  `Sleep wake disorders` + `Alcohol-related disorders` +
    `Depressive or anxiety disorders` + BENE_RACE_CD + AGE_CALC,
  method = "BFGS")

hrs_male <- hazard.msm(model_male); hrs_male <- lapply(hrs_male, hr_format)
hrs_female <- hazard.msm(model_female); hrs_female <- lapply(hrs_female, hr_format)

hrs_male_summary <- do.call(rbind, hrs_male)
hrs_female_summary <- do.call(rbind, hrs_female)
colnames(hrs_male_summary) <- colnames(hrs_female_summary) <- c("1-2","1-3","2-3")
rownames(hrs_male_summary) <- rownames(hrs_female_summary) <- c(names(data)[4:10], "Race", "Age (years)")

# print summary
hrs_male_summary <- as.data.frame(hrs_male_summary) %>%
  rownames_to_column(var = "Covariate") %>% gt()%>%
  tab_source_note(md("1 = Prediabetes, 
                     2 = Diabetes mellitus without complication,
                     3 = Diabetes mellitus with complication")) %>%
  tab_header(md("**Estimated Effect per Transition: Males**"))

hrs_female_summary <- as.data.frame(hrs_female_summary) %>%
  rownames_to_column(var = "Covariate") %>% gt()%>%
  tab_source_note(md("1 = Prediabetes, 
                     2 = Diabetes mellitus without complication,
                     3 = Diabetes mellitus with complication")) %>%
  tab_header(md("**Estimated Effect per Transition: Females**"))

hrs_male_summary
hrs_female_summary
gtsave(data = hrs_male_summary, filename =  "hrs_male_summary.html", path = here::here())
gtsave(data = hrs_female_summary, filename =  "hrs_female_summary.html", path = here::here())
```



```{r}
# transition probabilities
times <- c(0:12, seq(15, max(data$TIME_SINCE_PREDIAB)+3, 3)) #times to compute probs for...
transition_state1_m <- transition_state2_m <- transition_state1_f <- transition_state2_f <- c()

for (tm in times) {
  transition_state1_m <- rbind(transition_state1_m, as.vector(pmatrix.msm(model_male, t = tm)[1,]))
  transition_state2_m <- rbind(transition_state2_m, as.vector(pmatrix.msm(model_male, t = tm)[2,2:3]))
  
  transition_state1_f <- rbind(transition_state1_f, as.vector(pmatrix.msm(model_female, t = tm)[1,]))
  transition_state2_f <- rbind(transition_state2_f, as.vector(pmatrix.msm(model_female, t = tm)[2,2:3]))
}

colnames(transition_state1_m) <- colnames(transition_state1_f) <- c("State 1","State 2","State 3")
colnames(transition_state2_m) <- colnames(transition_state2_f) <- c("State 2","State 3")
rownames(transition_state1_m) <- rownames(transition_state2_m) <- rownames(transition_state1_f) <- rownames(transition_state2_f) <- times


# transition probability curves
transition_state1_curve <- 
  rbind(as.data.frame(transition_state1_m) %>% mutate(Sex = "Male") %>% rownames_to_column(var = "time"),
        as.data.frame(transition_state1_f) %>% mutate(Sex = "Female") %>% rownames_to_column(var = "time")) %>%
  mutate(trns_from = "State 1")

transition_state2_curve <- 
  rbind(as.data.frame(transition_state2_m) %>% mutate(Sex = "Male") %>% rownames_to_column(var = "time"),
        as.data.frame(transition_state2_f) %>% mutate(Sex = "Female") %>% rownames_to_column(var = "time")) %>%
  mutate(`State 1` = NA, trns_from = "State 2") %>%
  select(time, `State 1`, `State 2`, `State 3`, Sex, trns_from) 

transition_curve_A <- rbind(transition_state1_curve, transition_state2_curve) %>%
  mutate(time = as.numeric(time)) %>%
  pivot_longer(cols = starts_with("State"), names_to = "State", values_to = "prob") %>%
  mutate(State = stringr::str_sub(State, 7)) %>%
  ggplot(., aes(x = time, y = prob, color = State, linetype = Sex)) +
  geom_line() +
  facet_wrap(trns_from ~., nrow = 1) +
  labs(x="Time (months)", y="Probability") +
  theme_pubr()

#transition_curve_A
```



## Statify by Age groups
* Compare age-group adjusted transition probabilities (age<65 vs. age=65+)

```{r}
table(data$AGE_GRP)
```


```{r}
# separate msm for age<65 vs. age=65
model_yng <- msm(
  STATE ~ TIME_SINCE_PREDIAB, subject = BENE_ID,
  data = subset(data, AGE_GRP=="ADULT BELOW 65"), qmatrix = Q,
  covariates = ~ Obesity + `Socioeconomic or lifestyle factors` + Hypertension +
    `Heart disease or failure` +  `Sleep wake disorders` + `Alcohol-related disorders` +
    `Depressive or anxiety disorders` + SEX_IDENT_CD + BENE_RACE_CD,
  method = "BFGS")

model_old <- msm(
  STATE ~ TIME_SINCE_PREDIAB, subject = BENE_ID,
  data = subset(data, AGE_GRP=="ADULT 65+"), qmatrix = Q,
  covariates = ~ Obesity + `Socioeconomic or lifestyle factors` + Hypertension +
    `Heart disease or failure` +  `Sleep wake disorders` + `Alcohol-related disorders` +
    `Depressive or anxiety disorders` + SEX_IDENT_CD + BENE_RACE_CD,
  method = "BFGS")

hrs_yng <- hazard.msm(model_yng); hrs_yng <- lapply(hrs_yng, hr_format)
hrs_old <- hazard.msm(model_old); hrs_old <- lapply(hrs_old, hr_format)

hrs_yng_summary <- do.call(rbind, hrs_yng)
hrs_old_summary <- do.call(rbind, hrs_old)
colnames(hrs_yng_summary) <- colnames(hrs_old_summary) <- c("1-2","1-3","2-3")
rownames(hrs_yng_summary) <- rownames(hrs_old_summary) <- c(names(data)[4:10], "Sex", "Race")

# print summary
hrs_yng_summary <- as.data.frame(hrs_yng_summary) %>%
  rownames_to_column(var = "Covariate") %>% gt()%>%
  tab_source_note(md("1 = Prediabetes, 
                     2 = Diabetes mellitus without complication,
                     3 = Diabetes mellitus with complication")) %>%
  tab_header(md("**Estimated Effect per Transition: Adult Below 65**"))

hrs_old_summary <- as.data.frame(hrs_old_summary) %>%
  rownames_to_column(var = "Covariate") %>% gt()%>%
  tab_source_note(md("1 = Prediabetes, 
                     2 = Diabetes mellitus without complication,
                     3 = Diabetes mellitus with complication")) %>%
  tab_header(md("**Estimated Effect per Transition: Adult 65+**"))

hrs_yng_summary
hrs_old_summary
gtsave(data = hrs_yng_summary, filename =  "hrs_yng_summary.html", path = here::here())
gtsave(data = hrs_old_summary, filename =  "hrs_old_summary.html", path = here::here())
```



```{r}
# transition probabilities
times <- c(0:12, seq(15, max(data$TIME_SINCE_PREDIAB)+3, 3)) #times to compute probs for...
transition_state1_y <- transition_state2_y <- transition_state1_o <- transition_state2_o <- c()

for (tm in times) {
  transition_state1_y <- rbind(transition_state1_y, as.vector(pmatrix.msm(model_yng, t = tm)[1,]))
  transition_state2_y <- rbind(transition_state2_y, as.vector(pmatrix.msm(model_yng, t = tm)[2,2:3]))
  
  transition_state1_o <- rbind(transition_state1_o, as.vector(pmatrix.msm(model_old, t = tm)[1,]))
  transition_state2_o <- rbind(transition_state2_o, as.vector(pmatrix.msm(model_old, t = tm)[2,2:3]))
}

colnames(transition_state1_y) <- colnames(transition_state1_o) <- c("State 1","State 2","State 3")
colnames(transition_state2_y) <- colnames(transition_state2_o) <- c("State 2","State 3")
rownames(transition_state1_y) <- rownames(transition_state2_y) <- rownames(transition_state1_o) <- rownames(transition_state2_o) <- times


# transition probability curves
transition_state1_curve <- 
  rbind(as.data.frame(transition_state1_y) %>% mutate(Age = "Adult 65-") %>% rownames_to_column(var = "time"),
        as.data.frame(transition_state1_o) %>% mutate(Age = "Adult 65+") %>% rownames_to_column(var = "time")) %>%
  mutate(trns_from = "State 1")

transition_state2_curve <- 
  rbind(as.data.frame(transition_state2_y) %>% mutate(Age = "Adult 65-") %>% rownames_to_column(var = "time"),
        as.data.frame(transition_state2_o) %>% mutate(Age = "Adult 65+") %>% rownames_to_column(var = "time")) %>%
  mutate(`State 1` = NA, trns_from = "State 2") %>%
  select(time, `State 1`, `State 2`, `State 3`, Age, trns_from) 

transition_curve_B <- rbind(transition_state1_curve, transition_state2_curve) %>%
  mutate(time = as.numeric(time)) %>%
  pivot_longer(cols = starts_with("State"), names_to = "State", values_to = "prob") %>%
  mutate(State = stringr::str_sub(State, 7)) %>%
  ggplot(., aes(x = time, y = prob, color = State, linetype = Age)) +
  geom_line() +
  facet_wrap(trns_from ~., nrow = 1) +
  labs(x="Time (months)", y="Probability") +
  theme_pubr()

#transition_curve_B
```



```{r}
# plot
# combined_plot <- plot_grid(transition_curve_A, transition_curve_B, labels = c("A", "B"), ncol = 1)
# combined_plot

transition_curve_A
transition_curve_B
```
